requirements.txt
keras
tensorflow-gpu
joblib
tqdm
h5py
